Bug on component attribute to html attribute
